policy

+++ app/Http/Controllers/Designs/CommentController.php
@@ -42,7 +42,7 @@ class CommentController extends Controller
     public function update(Request $request , $id)
     {
         $comment = $this->comments->find($id);
-        // $this->authorize('update', $comment);
+        $this->authorize('update', $comment);
 
         $this->validate($request, [
             'body' => ['required']
@@ -63,7 +63,7 @@ class CommentController extends Controller
     public function destroy($id)
     {
         $comment = $this->comments->find($id);
-        // $this->authorize('delete',$comment);
+        $this->authorize('delete',$comment);
         $this->comments->delete($id);
         return response()->json(
             [
